Validate known unexpected nodes
This commit adds a new `UnexpectedNode` type that will wrap unexpected
nodes that are parsed from invalid Ruby code. This will make it easier
for clients of Prism to understand that a node was unexpected.

For example:

```
module def foo end
```

This previously parsed `ModuleNode` to have a `constant_path` field with
a `DefNode`. Clients currently need to handle this error recovery case
by checking the allowed node types. With this change, they will only
need to check that it is not a `MissingNode` or `UnexpectedNode`.

The `UnexpectedNode` has a `child` field containing the original node so
clients can choose to pull it out and do something with it. This will
obviously depend on the use-case, but Prism can make it easier to
understand the node was unexpected and clients that don't care can
potentially just ignore it.