wip! Validate all expected nodes
The previous commit added validation to fields that already have known
error recovery states (as per `config.yml`). It's possible that those
are the only nodes that can have unexpected nodes, but this commit adds
validation to _all_ fields that have expected node types. I'm not sure
if this is necessary or not.